Ensure Matrix alias updated first
If we are bridging a Matrix room that the bot
does not have control over, the API calls to
update the name or topic will fail and skip
over all of the alias updating that comes later on.

We can control the #*:gitter.im aliases no matter what
so we are updating those first, then trying to update
the rest if we can.