Update graceful-fs (gulp dependency) to be compatible in Node v14
Fix `ReferenceError: primordials is not defined` when running
the webapp with Node v14.

Solved via https://stackoverflow.com/questions/55921442/how-to-fix-referenceerror-primordials-is-not-defined-in-node/58394828#58394828

Created a script to iterate our whole package-lock.json and remove all of the
resolved entries of `graceful-fs`. Then next time we `npm i`, they resolve to the
latest 4.2.6! Plus we get the benefit of using a single version for them all!

Part of https://gitlab.com/gitterHQ/webapp/-/issues/2735