Avoid join->send race condition by joining long before the message send
Fix https://gitlab.com/gitterHQ/webapp/-/issues/2770

If we join on Matrix when the user joins the room on Gitter,
there will be a significant gap of time to allow the join event
to federate before the user sends a message.

The [underlying Synapse race condition](https://github.com/matrix-org/synapse/issues/10066)
could still be triggered if the user was fast enough to join and send a message right away
(plus the variable amount of lag between each piece in the chain).