Make the "Reply" option ("Reply in Thread", "Start a Thread") always use "Reply in Thread" for consistency
This commit changes the "..." menu to always use "Reply in Thread" as the text for replying to a Gitter
message to provide a consistent option to look for when desiring to reply to something.  Speaking for
myself, I'm always looking for an option called "Reply" (coming from Slack, email, etc.) and often overlook
the "Start a New Thread" option, snapping instead to "Report" based on the common "Rep" prefix before
realizing that I've accidentally taken the wrong action.