Don't try to lookup by `virtualUser`/`parentId` at the same time which can be very slow
Follow-up to https://gitlab.com/gitterHQ/webapp/-/merge_requests/2313

Part of https://gitlab.com/gitterHQ/webapp/-/issues/2609

---

Why and what is it slow?

Our query to iterate over the main messages in the room looks like the following and can be very slow. I'm noticing this particularly in rooms where a single message was sent in the live room and was already bridged to Matrix. My guess is that this is because it can't find a message ID that is less than the live message ID we want to stop at, it keeps paginating through the whole history of messages. This kinda doesn't make sense though as it works fine for rooms with more messages, and from the query plan explanation I don't really see any difference to the stages (sort, input, filter) or in the `executionStats` (`totalKeysExamined`/`totalDocsExamined`). Or it's just that we don't have a compound index for `virtualUser`/`parentId` :shrug: