Fix flaw in room resumption code skipping room it should have resumed from
See https://gitlab.com/gitterHQ/webapp/-/issues/2609#note_1239203705

Fixing bug in our room resumption code where our resume position checkpoint
file tracks the oldest room ID we are currently processing. But then we're
using it as `$gt: resumeFromGitterRoomId` which goes to the room after the
room we are supposed to resume from. We should be using `$gte:
resumeFromGitterRoomId`. This doesn't seem like it would account for so much
difference as we didn't restart the script that many times.

We now better disambiguate "resume" from "lastThingProcessed" type language
to better distinguish when to use `$gt` vs `$gte`.