Persist room ID's that failed to import along the way so we can review
Before we would only persist all the failures at the end but that means we
can't debug anything while it's still running. And even if we can see the
shell terminal output for the running import process, sometimes it
disconnects and we lose the insight.

Follow-up to https://gitlab.com/gitterHQ/webapp/-/merge_requests/2313

Part of https://gitlab.com/gitterHQ/webapp/-/issues/2609